Replacing answers with example domain
answers.openmrs.org is now a redirect so we can't get a valid response on the necessary URL. In order for the site to be functional, we put this hack into production. This is a TEMPORARY fix and we need to get rid of this dependency, but this hack is necessary for production to work now that the site is shut down.